requirements.txt
# Pip Requirements File
shodan 
python-dotenv
requests
pymongo
flask
python-nmap
fastapi
ping3